test(devnet): separate shared sweep topology preflight
